Revert "Exhibition"
