+++ src/components/ShareDetailComponent/ShareDetailLeader.scss
@@ -69,7 +69,7 @@
     }
     &-AddImage_Bundle {
       margin-top: 5%;
-      .Addimage_box {
+      .Addimage_box_1 {
         margin-top: 1%;
         width: 30%;
         height: 30vh;
