+++ src/components/ShareDetailComponent/ShareDetailLeader.tsx
@@ -1,11 +1,11 @@
 import { useLocation } from "react-router-dom";
 import useModalStore from "../../store/useModalState";
-import { useState } from "react";
+import { useRef, useState } from "react";
 import Header from "../header/Header";
 import "./ShareDetailLeader.scss";
 import RootModal from "../Modal/RootModal/RootModal";
 import nouserImage from "../../assets/img/noimage.jpg";
-import useImageUpload from "../../hooks/useImageUpload";
+import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 
 interface ShareDetailLeaderProps {
   userId: number;
@@ -28,13 +28,63 @@ const ShareDetailLeader = () => {
   const [modalName, setModalName] = useState("");
   const [formData, setFormData] = useState<ShareDetailLeaderProps>(initialData);
   const [projectLink, setProjectLink] = useState("");
-  const { imageUrl, handleFileChange, handleUpload, postUplodFileUrl } =
-    useImageUpload();
 
-  const saveShareDetail = async () => {
-    const s3ImageUrl = await handleUpload();
-    postUplodFileUrl(s3ImageUrl);
+  const [selectedFile, setSelectedFile] = useState<File | null>(null); //FIXME: 추가됨
+  const [imageUrl, setImageUrl] = useState<string | null>(null);
+  const s3ImageUrl = useRef("");
+
+  const s3Client = new S3Client({
+    region: "ap-northeast-2",
+    credentials: {
+      accessKeyId: import.meta.env.VITE_SECRET_KEY,
+      secretAccessKey: import.meta.env.VITE_SECRET_ACCESS_KEY,
+    },
+  });
+
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0] || null;
+    if (file) {
+      setSelectedFile(file);
+      const objectUrl = URL.createObjectURL(file);
+      setImageUrl(objectUrl);
+      console.log("⭐미리보기:", imageUrl);
+    }
   };
+
+  const handleUpload = async () => {
+    if (!selectedFile) {
+      alert("파일을 선택해주세요");
+      return;
+    }
+    try {
+      const uploadParams = {
+        Bucket: import.meta.env.VITE_IMG_S3,
+        Key: `folder${selectedFile.name}`, // S3에 저장될 경로와 파일명
+        Body: selectedFile,
+        ContentType: selectedFile.type,
+      };
+      const command = new PutObjectCommand(uploadParams);
+      const response = await s3Client.send(command);
+      console.log("File upload successfully:", response);
+
+      const uploadedUrl = `https://${
+        import.meta.env.VITE_IMG_S3
+      }.s3.ap-northeast-2.amazonaws.com/folder${encodeURIComponent(
+        selectedFile.name
+      )}`;
+      s3ImageUrl.current = uploadedUrl;
+    } catch (error) {
+      console.error("Error uploading file: ", error);
+    }
+  };
+  const postUplodFileUrl = () => {
+    console.log("preImgUrl", s3ImageUrl.current);
+  };
+  const saveShareDetail = () => {
+    handleUpload();
+    postUplodFileUrl();
+  };
+
   /** 입력 값 변경 핸들러 */
   const handleInputChange = (
     e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
@@ -50,6 +100,10 @@ const ShareDetailLeader = () => {
   ) => {
     setProjectLink(e.target.value);
   };
+
+  // TODO: 1. 프론트에서 s3로 직접 이미지 업로드
+  //2.이미지 링크 받아와서 str로 백엔드에게 전달
+  /** 멤버 초대 모달 열기 */
   const invalidateInstance = () => {
     setModalName("shareInviteModal");
     openModal();
@@ -139,6 +193,12 @@ const ShareDetailLeader = () => {
               </div>
             </div>
           </div>
+          <img
+            className="abc"
+            src="https://lymming-img.s3.amazonaws.com/folder/외주요청.png
+"
+            alt=""
+          />
           <div className="ShareDetailLeader-Footer_BtnWrapper">
             <div className="saveBtn" onClick={saveShareDetail}>
               저장하기
