+++ src/store/useLoginStore.ts
@@ -21,7 +21,7 @@ interface InfoState {
 
 // 로그인 data
 export const useLoginStore = create<LoginState>((set) => ({
-  login: true,
+  login: false,
   count: 1,
   isOpen: false,
   //닉네임 중복체크
@@ -44,7 +44,7 @@ export const useInfoStore = create<InfoState>((set) => ({
     interests: [],
     job: "",
     loginType: "",
-    nickname: "노기훈",
+    nickname: "",
     position: "",
     stack: [],
     temperature: 0,
